Port Wolfenstein 3D to macOS Apple Silicon with SDL2 (#2)
* Port Wolfenstein 3D to macOS Apple Silicon with SDL2

Complete rewrite of id Software's original 1992 DOS Wolfenstein 3D source for native macOS arm64 execution via SDL2. Replaced all DOS/Borland-specific code with modern POSIX equivalents:

- Platform abstraction: SDL2 for graphics, audio, input, and timers
- Memory management: Standard malloc/free replacing EMS/XMS
- Video system: Flat SDL2 framebuffer with indexed color to RGBA conversion
- Audio: SDL2 mixer for sound/music (OPL2 emulator, PCM samples)
- Input: SDL2 event polling replacing DOS interrupt handlers
- Assembly elimination: All x86 ASM replaced with C (raycaster, sprite scaling, etc.)
- Type fixes: 16-bit DOS int assumptions corrected for 64-bit systems
- Build system: Makefile for clang targeting arm64 with -std=c99

Compiles cleanly to native Mach-O arm64 executable (288KB, 193 warnings, 0 errors).
Requires Wolfenstein 3D data files (VSWAP.WL6, etc.) and SDL2 from Homebrew.


* Fix critical runtime bugs and clean up build artifacts

- Fix SIGSEGV crash: move VL_Startup() before SignonScreen() so SDL
  framebuffer is allocated before any rendering code touches it
- Fix screenofs calculation: use flat 320-pixel-wide offsets instead of
  VGA Mode X 80-byte-wide values, fixing 3D viewport positioning
- Fix data file case sensitivity: add case-insensitive file open helpers
  in id_ca.c, id_pm.c, and wl_menu.c for macOS case-sensitive volumes
- Fix CPU-burning busy-wait loops: add SDL_Delay(1) in CalcTics and
  PollControls so the game yields CPU while waiting for timer ticks
- Fix VL_FadeIn cumulative rounding: save starting palette and
  interpolate from fixed values instead of modifying in-place each step
- Add SDL_Quit() call in VL_Shutdown for proper cleanup
- Add .gitignore and remove obj/ and wolf3d binary from tracking


* Fix 64-bit type size mismatches in data file structures

On arm64 macOS, `long` is 8 bytes and `int`/`unsigned` are 4 bytes, but
the Wolf3D data files store these values as 2-byte or 4-byte fields
(matching the original 16-bit DOS `int` = 2 bytes, `long` = 4 bytes).

Reading file data into native-width types caused garbage values and
immediate crashes (MM_GetPtr called with size 0xaaaaa95f0000206c).

Fixed by using explicit fixed-width types for all file-format structs:
- grstarts/audiostarts: long* -> int32_t* (file offsets are 32-bit)
- mapfiletype.headeroffsets: long[] -> int32_t[]
- mapfiletype.RLEWtag: unsigned -> uint16_t
- maptype.planestart: long[] -> int32_t[]
- maptype fields: unsigned -> uint16_t
- pictabletype: int -> int16_t
- CAL_ExpandGrChunk expanded size: *(long*) -> *(int32_t*)


* Fix CheckForEpisodes falling through to WL1 check after WL6 found

The preprocessor logic had a bug where finding VSWAP.WL6 would set the
extension correctly, but code would then fall through past the #ifdef
SPEAR / #else block and unconditionally check for VSWAP.WL1, quitting
with "NO WOLFENSTEIN 3-D DATA FILES to be found!" when WL1 wasn't
present. Added a guard so the WL1 fallback only runs if no extension
was already set by the WL6/WL3 check.


* Fix data file discovery when running from any directory

Use _NSGetExecutablePath() to chdir to the executable's directory at
startup so data files are found regardless of the user's working
directory. Also remove unused boolean_t typedef that conflicted with
macOS system headers.


* Fix bus error: US_Print/US_CPrint wrote to read-only string literals

On modern systems, string literals are in read-only memory. The original
code temporarily null-terminated substrings in place, which causes SIGBUS
on ARM64. Fixed by copying each line to a local buffer before processing.


* Extract real game palette and signon screen from original OBJ files

The game was showing a black screen because gamepal[768] was all zeros.
In the original DOS build, the palette and signon screen were linked from
GAMEPAL.OBJ and SIGNON.OBJ binary blobs. This extracts the actual 768-byte
VGA palette and 64000-byte signon screen from those OMF object files and
embeds them in signon.c. Also fixes duplicate gamepal definition in id_vh.c.


* Add VL_Present() calls during startup for SDL2 rendering

In DOS, writing to VGA memory was immediately visible on screen. In SDL2,
the framebuffer must be explicitly presented via SDL_RenderPresent. Added
VL_Present() calls after SignonScreen draws the signon image and in
FinishSignon before waiting for user input, so the startup screens are
actually visible instead of showing a black window.


* Fix garbled screens: separate planar and linear rendering paths

VL_MungePic converts planar data to linear pixel order, but VL_MemToScreen
was then trying to de-interleave again, double-processing the data.

Split into two functions:
- VL_MemToScreen: straight linear copy (for data already through MungePic)
- VL_PlanarToScreen: de-interleaves plane-sequential to linear (for cached
  graphics from VGAGRAPH that haven't been munged)

Also added crash signal handler for debugging.


* Port Wolfenstein 3D to macOS Apple Silicon with SDL2

Complete port of id Software's 1992 DOS Wolfenstein 3D source to native macOS arm64 using SDL2.

Major changes:
- Replaced all x86 assembly (10 ASM files + inline asm) with portable C implementations
- Rewrote VGA graphics system: planar Mode X framebuffer → SDL2 flat framebuffer with proper de-interleaving
- Replaced DOS audio/timer interrupts → SDL2 mixer and timer events
- Replaced DOS keyboard/mouse ISRs → SDL2 event polling
- Replaced EMS/XMS memory management with standard malloc/free
- Fixed all 64-bit type size mismatches (long, int, unsigned, far pointers)
- Fixed struct alignment issues (mapfiletype, maptype) with __attribute__((packed))
- Fixed data format issues (Huffman, Carmack, RLEW, sprite commands all use 16-bit words)
- Renders loading screens, title screen, and menu correctly
- Game initializes, loads levels, and executes raycasting/AI

Added wl_compat.h compatibility header mapping Borland Turbo C++ and DOS-specific types/macros to modern POSIX/C99 equivalents.

Compiles to 288KB native Mach-O arm64 binary with included Makefile targeting clang.


---------

+++ src/wl_def.h
@@ -1071,8 +1071,8 @@ typedef struct
 
 typedef struct
 {
-	unsigned	leftpix,rightpix;
-	unsigned	dataofs[64];
+	uint16_t	leftpix,rightpix;
+	uint16_t	dataofs[64];
 // table data after dataofs[rightpix-leftpix+1]
 }	t_compshape;
